# section-04/lab05/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Purchase Approval API",
        "summary": "구매 요청 생성·조회 및 Workflow 승인 결과 저장",
        "description": "watsonx Orchestrate가 결정한 구매 승인/반려 결과를 MySQL에 저장하는 purchase app입니다. HTTP 처리, 업무 규칙, DB 처리, 연결 관리를 각각 routes, services, repositories, db 모듈로 분리합니다.\n\n**호출 순서:** POST /api/purchase-requests로 생성 → 응답 id 보관 → Workflow에서 결정 → PATCH /api/purchase-requests/{request_id}/status로 결과 저장 → GET으로 현재 상태 확인.\n\n현재 상태는 PENDING / APPROVED / REJECTED입니다. 이 앱에는 견적·품목 조회·발주·결제 API가 없습니다. purchase_type, requester, approver_role은 전달받아 저장하는 문자열이며 서버 인증/역할 검사를 의미하지 않습니다. 현재 실습 앱에는 애플리케이션 인증이 없습니다.\n\n구매 총액은 Decimal이며 응답 JSON에서는 문자열로 반환됩니다. 날짜·시각에는 timezone 정보가 없습니다. 오류는 FastAPI 형식을 사용하며 404/409는 detail 문자열, 입력 검증 422는 detail 배열입니다. 처리되지 않은 DB 오류에는 JSON 오류 변환 핸들러가 없습니다.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://nexweb.ddnsgeek.com/purchase",
            "description": "HTTPS API through Nginx"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "구매 가능한 물품 목록 조회",
                "description": "구매 요청에 사용할 수 있는 활성 물품 목록을 SKU, 물품명, 단가와 함께 반환합니다. For each 노드에서 items의 sku와 비교할 때 사용할 수 있습니다.",
                "operationId": "list_products",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResponse"
                                    },
                                    "type": "array",
                                    "title": "Response List Products"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "구매 물품 등록",
                "description": "구매 요청에 사용할 물품을 SKU, 물품명, 단가와 함께 등록합니다. SKU는 대문자로 정규화되며 이미 등록된 SKU는 다시 등록할 수 없습니다.",
                "operationId": "create_product",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "이미 등록된 SKU입니다."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{sku}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "SKU로 물품 및 단가 조회",
                "description": "지정한 SKU의 활성 물품을 조회하고 1개 기준 단가(unit_price)를 반환합니다. SKU는 대소문자를 구분하지 않으며 서버에서 대문자로 정규화합니다.",
                "operationId": "get_product_by_sku",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 50,
                            "description": "조회할 물품 SKU",
                            "examples": [
                                "NOTEBOOK",
                                "MONITOR",
                                "KEYBOARD"
                            ],
                            "title": "Sku"
                        },
                        "description": "조회할 물품 SKU"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "해당 SKU의 활성 물품이 없습니다."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ProductCreate": {
                "properties": {
                    "sku": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 1,
                        "title": "Sku",
                        "description": "물품을 식별하는 SKU. 서버에서 대문자로 정규화합니다.",
                        "examples": [
                            "MOUSE"
                        ]
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 1,
                        "title": "Name",
                        "description": "물품명",
                        "examples": [
                            "마우스"
                        ]
                    },
                    "unit_price": {
                        "anyOf": [
                            {
                                "type": "number",
                                "exclusiveMinimum": 0.0
                            },
                            {
                                "type": "string",
                                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                            }
                        ],
                        "title": "Unit Price",
                        "description": "물품 1개의 단가",
                        "examples": [
                            50000
                        ]
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description",
                        "description": "물품 설명"
                    },
                    "active": {
                        "type": "boolean",
                        "title": "Active",
                        "description": "등록 즉시 구매 가능한 물품인지 여부",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "sku",
                    "name",
                    "unit_price"
                ],
                "title": "ProductCreate",
                "description": "물품 등록 API의 요청 모델입니다."
            },
            "ProductResponse": {
                "properties": {
                    "sku": {
                        "type": "string",
                        "title": "Sku",
                        "description": "물품을 식별하는 대문자 SKU",
                        "examples": [
                            "NOTEBOOK"
                        ]
                    },
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "물품명",
                        "examples": [
                            "노트북"
                        ]
                    },
                    "unit_price": {
                        "type": "string",
                        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                        "title": "Unit Price",
                        "description": "물품 1개의 단가",
                        "examples": [
                            1200000
                        ]
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description",
                        "description": "물품 설명"
                    },
                    "active": {
                        "type": "boolean",
                        "title": "Active",
                        "description": "현재 구매 가능한 물품인지 여부",
                        "examples": [
                            true
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "sku",
                    "name",
                    "unit_price",
                    "active"
                ],
                "title": "ProductResponse",
                "description": "물품 조회 API의 응답 모델입니다."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}